Delete pyproject.toml
pyproject.toml was previously used for the validation service. This has been configured to run without python thus this file is no longer needed.

Dependabot has found an issue with this file "Jinja vulnerable to HTML attribute injection when passing user input as keys to xmlattr filter" so best to delete.